% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{adjectives}
\alias{adjectives}
\title{Adjectives for generating journalnames}
\format{
An object of class \code{data.frame} with 5137 rows and 2 columns.
}
\usage{
adjectives
}
\description{
A selection of adjectives for generating journal names
}
\keyword{datasets}
