% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folketingsvalg}
\alias{folketingsvalg}
\title{Folketingsvalg siden 1953}
\format{
A date vector with 24 dates
}
\source{
https://www.ft.dk/-/media/sites/ft/pdf/folkestyret/valg-og-afstemninger/folketingsvalg-fra-1953.ashx
}
\usage{
folketingsvalg
}
\description{
Dates for Danish parlamentary elections since 1953-09-22
}
\keyword{datasets}
