% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_journals.R
\name{generate_journals}
\alias{generate_journals}
\title{generate_journals()}
\usage{
generate_journals(n = 10, subjects = T)
}
\arguments{
\item{n}{number of journals returned}

\item{subjects}{logical value indicating wether the subject of the journal should be returned}
}
\description{
generate_journals returns a tibble with fictional names of scientific journals
}
\details{
details here - vi skal huske at kreditere lloyd
}
\references{
L. James, ESG, (2021), GitHub repository, https://github.com/ltjames/ESG
}
