% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_sciences.R
\name{generate_sciences}
\alias{generate_sciences}
\title{generate_sciences()}
\usage{
generate_sciences(n = 10L, subjects = T)
}
\arguments{
\item{n}{number of scientific disciplines returned}

\item{subjects}{logical value indicating wether the major subject of the discipline should be returned}
}
\description{
generate_sciences returns a tibble with fictional scientific disciplines
}
\details{
details here - vi skal huske at kreditere lloyd
}
\references{
L. James, ESG, (2021), GitHub repository, https://github.com/ltjames/ESG
}
