% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_spiral.R
\name{geom_spiral}
\alias{geom_spiral}
\title{geom_spiral}
\usage{
geom_spiral(
  mapping = NULL,
  data = NULL,
  geom = "path",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  n = 12,
  ...
)
}
\arguments{
\item{param}{1}
}
\description{
geom_spiral laver et spiral plot.
}
\details{
details here
}
