% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hp_characters}
\alias{hp_characters}
\title{Characters from Harry Potter}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 90 rows and 3 columns.
}
\usage{
hp_characters
}
\description{
A selection of characters from Harry Potter
}
\keyword{datasets}
