% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{journal_patterns}
\alias{journal_patterns}
\title{Patterns for journal generator}
\format{
An object of class \code{data.frame} with 13 rows and 1 columns.
}
\usage{
journal_patterns
}
\description{
Patterns for journal generator
}
\keyword{datasets}
