% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kursuskalender.R
\name{kursuskalender}
\alias{kursuskalender}
\alias{coursecalendar}
\title{Kursuskalender}
\usage{
kursuskalender()

coursecalendar()
}
\description{
kursuskalender() opens the KUB Datalab course calendar in a browser
}
\details{
details here
}
