% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%nin\%}
\alias{\%nin\%}
\title{Negated %in%}
\usage{
x \%nin\% table
}
\arguments{
\item{param}{2}
}
\description{
%nin% returns a logical vector indicating if there is a match or not for its left operand
}
\details{
details here
}
