% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sciences}
\alias{sciences}
\title{Sciences for journal generator}
\format{
An object of class \code{data.frame} with 692 rows and 1 columns.
}
\usage{
sciences
}
\description{
Sciences for journal generator
}
\keyword{datasets}
