% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_mode.R
\name{stat_mode}
\alias{stat_mode}
\title{Mode}
\usage{
stat_mode(x)
}
\arguments{
\item{x}{den vektor hvor vi vil have typetallet}
}
\description{
stat_mode returns the mode (most common value) in a vector
}
\details{
details here
}
