% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatSpiral.R
\docType{data}
\name{StatSpiral}
\alias{StatSpiral}
\title{StatSpiral}
\format{
An object of class \code{StatSpiral} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatSpiral
}
\description{
StatSpiral returnerer det vi skal bruge for at lave geom_spiral
}
\details{
details here
}
\keyword{datasets}
