% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_kubdatalab.R
\name{theme_kubdatalab}
\alias{theme_kubdatalab}
\title{KUBDatalab theme}
\usage{
theme_kubdatalab(
  text.font = NULL,
  title.font = NULL,
  legend.font = NULL,
  title.size = 18,
  text.size = 14,
  subtitle.size = 12,
  axis.title.size = 14,
  axis.text.size = 10,
  legend.title.size = 10,
  legend.text.size = 9,
  title.color = "#000000",
  subtitle.color = "#000000",
  text.color = "#000000",
  axis.title.color = "#000000",
  axis.text.color = "#000000",
  panel.grid.major.color = "#000000",
  panel.grid.minor.color = "grey",
  legend.title.color = "#000000",
  legend.text.color = "#000000",
  legend.position = "bottom",
  ticks = FALSE
)
}
\arguments{
\item{text.font}{text font, Default: NULL}

\item{title.font}{title font, Default: NULL}

\item{legend.font}{legend font, Default: NULL}

\item{title.size}{title font size, Default: 18}

\item{text.size}{text font size, Default: 14}

\item{subtitle.size}{subtitle font size, Default: 12}

\item{axis.title.size}{axis title font size, Default: 14}

\item{axis.text.size}{axis text font size, Default: 10}

\item{legend.title.size}{legend title font size, Default: 10}

\item{legend.text.size}{legend text font size, Default: 9}

\item{title.color}{title color, Default: "#FFD235"}

\item{subtitle.color}{subtitle.color, Default: "#fee8c8"}

\item{text.color}{text color, Default: "#fee8c8"}

\item{axis.title.color}{axis title color, Default: "#fee8c8"}

\item{axis.text.color}{axis text color, Default: "#fee8c8"}

\item{panel.grid.major.color}{en farveting}

\item{panel.grid.minor.color}{en anden farveting}

\item{legend.title.color}{legend title color, Default: "#ffffff"}

\item{legend.text.color}{legend text color, Default: "#ffffff"}

\item{legend.position}{legend position, Default: "bottom"}

\item{ticks}{add axis ticks, Default: FALSE}
}
\description{
KUBDatalab theme
}
\details{
Inspired by tvthemes by Ryo Nakagawara
}
\examples{
library(ggplot2)

ggplot(airquality, aes(x = Day, y = Temp,
     group = as.factor(Month), color = as.factor(Month))) +
     geom_point(size = 2.5) +
     scale_color_ucph() +
     theme_kubdatalab()
}
\seealso{
[ggplot2::theme]
}
