% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{folketingsvalg}
\alias{folketingsvalg}
\title{Folketingsvalg siden 1953}
\format{
A date vector containing dates for danish parlamentary elections since 1953
}
\usage{
folketingsvalg
}
\description{
Folketingsvalg siden 1953
}
\keyword{datasets}
